examples/resume_agent/pyproject.toml
[project]
name = "resume-agent"
version = "0.1.0"
description = "Voice AI agent for recruiter calls — Syrin + Pipecat"
requires-python = ">=3.11"
dependencies = [
    "syrin[knowledge]",
    "pipecat-ai[deepgram,elevenlabs]>=0.0.90",
    "python-dotenv>=1.0",
]

[tool.uv]
dev-dependencies = []
[tool.uv.sources]
syrin = { path = "../..", editable = true }
